requirements.txt
Django==3.0.3
djangorestframework==3.11.0

requests==2.22.0
PyJWT==1.7.1
psycopg2

asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.8
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
